refactor: Popup 컴포넌트 데이터 넘기는 작업 진행, 팝업 내 항목 클릭 시 팝업 닫히지 않는 현상 해결
