+++ FE/src/components/FilterBar/index.tsx
@@ -3,27 +3,37 @@ import { css } from '@emotion/react';
 import * as S from './style';
 
 import I from '@components/Icons';
+import Popup from '@components/Popup';
 import Contents from '@components/Popup/Contents';
-import Popup from '@components/Popup/Popup';
 import { IPopupData } from '@components/Popup/type';
 import useComponentVisible from '@hooks/useComponentVisible.jsx';
+import { useSearch } from '@hooks/useSearch';
 
 export default function FilterBar() {
   const issueFilterData = {
     info: [
-      { id: 1, status: 'open', name: '열린 이슈' },
-      { id: 2, status: 'mine', name: '내가 작성한 이슈' },
-      { id: 3, status: 'assignedToMe', name: '나에게 할당된 이슈' },
-      { id: 4, status: 'comment', name: '내가 댓글을 남긴 이슈' },
-      { id: 5, status: 'close', name: '닫힌 이슈' },
+      { id: 1, status: 'is:open', name: '열린이슈' },
+      { id: 2, status: 'mine:@me', name: '내가작성한이슈' },
+      { id: 3, status: 'assignedToMe:@me', name: '나에게할당된이슈' },
+      { id: 4, status: 'comment:@me', name: '내가댓글을남긴이슈' },
+      { id: 5, status: 'is:close', name: '닫힌이슈' },
     ],
   };
 
   const { ref, isComponentVisible, setIsComponentVisible } =
     useComponentVisible(false);
 
+  // TODO 차후에 수정필요 replace 함수 대신 다른 함수로 작업
+  const { replace } = useSearch('q', 'is:open');
+
+  // TODO handleOnFilterPopup 내에서 setIsComponentVisible 함수 2번씩 렌더링 되는 현상 원인 파악 필요.
   const handleOnFilterPopup = () => {
-    setIsComponentVisible(true);
+    setIsComponentVisible(!isComponentVisible);
+  };
+
+  const handleItemClick = (popupData: IPopupData) => {
+    replace(popupData.status, popupData.name);
+    setIsComponentVisible(false);
   };
 
   return (
@@ -46,6 +56,7 @@ export default function FilterBar() {
                   key={`popup-${popupData.id}`}
                   label="이슈"
                   popupData={popupData}
+                  handleItemClick={handleItemClick}
                 />
               ))}
             </Popup>
