+++ FE/src/components/Popup/type.ts
@@ -1,7 +1,12 @@
+/**
+ * TODO
+ * 백엔드 API 설계까지 임시적으로 사용하는 Popup 더미데이터
+ *  */
+
 export interface IPopupData {
   id: number;
   imageUrl?: string;
-  status?: string;
+  status?: any;
   color?: string;
   name: string;
 }
